grandpa: move Error type to its own file
